crm-py-antigravity/pyproject.toml
[project]
name = "crm-py-antigravity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
]
